Mates were still overlooked. I debugged and finally figured out it still had to do with the int:16 stuff. I commit a much better solution now, tested it, and it works! Should we make move_t's value an int16_t too?

+++ trunk/inc/search_mtdf.h
@@ -52,7 +52,8 @@ public:
 private:
 	void iterate(int s);
 	move_t mtdf(int depth, int guess = 0);
-	move_t minimax(int depth, int shallowness = 0, int alpha = -INFINITY, int beta = +INFINITY);
+	move_t minimax(int depth, int shallowness = 0, 
+		int16_t alpha = -INFINITY, int16_t beta = +INFINITY);
 };
 
 #endif
